fix: Load with default curreny
When we use a URL with a default currency (for example ?moneda=USD) we
need to set the current currency on the store, this is done in the
LatestExchange component, on mount it check the url and set the store
state, but, because a bug in the store, the current currency wasn't
updated, so we were loading the correct currency, but we are not showing
it becaues the current was current.

This commit fixes the condition so when we load the requested currency
the current currency is updated.