feat: Add flyway
This commit adds flyway to handle database migrations.

The problem with the previous approach is that some places closes
branches a complex 'list' diff is required to get the current version of
the branches.

This can be easily done with SQL because we know the exact state of the
data before, in a java approach, we will need to make sure that from any
state we can go to the desired one.

This will make somewhat more verbose to add new branches, but I think
that in the long run it will be easier to maintain.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>